requirements.txt
# core ml libraries
torch>=1.12.0
numpy>=1.21.0
gymnasium>=0.29.0

# rl training
stable-baselines3>=1.7.0
tensorboard>=2.8.0

# data handling
pyyaml>=6.0
pandas>=1.4.0
matplotlib>=3.5.0
seaborn>=0.11.0

# utilities
tqdm>=4.64.0
wandb>=0.13.0  # optional for cloud logging
jupyter>=1.0.0

# testing
pytest>=7.0.0
pytest-cov>=4.0.0 